backend/pyproject.toml
[project]
name = "grocery-store-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "mysql-connector-python>=9.3.0",
    "python-dotenv>=1.1.1",
]
